Break up <Editor /> tests into smaller files
